guides/snowflake_mirroring/snowpark_native/mirror_pyiceberg/requirements.txt
snowflake-snowpark-python
pydantic-core==2.16.2